ECO-313: Refactor communication framework.
