+++ explorer/plugin/src/content/inject.ts
@@ -1,37 +1,10 @@
-import { Rpc } from '../lib/rpc/rpc';
-import { Request } from '../lib/rpc/tunnel';
-
-// A unique message ID that is used to ensure responses are sent to the correct requests
-let _messageId = 0;
-
-let generateNewMessageId = () => ++_messageId;
-
-const PAGE_ID = Math.random() * Number.MAX_SAFE_INTEGER;
+import { registerClient } from '../lib/rpc/client';
 
 class CasperLabsPluginHelper {
-  private rpc: Rpc;
+  private readonly call: <RESULT>(method: string, ...params: any[]) => Promise<RESULT>;
 
   constructor() {
-    this.rpc = new Rpc({
-      source: 'page',
-      destination: 'background',
-      logMessages: true,
-      postMessage: (msg: Request) => {
-        return new Promise((resolve, reject) => {
-          const msgId = generateNewMessageId();
-          window.postMessage({ type: 'request', pageId: PAGE_ID, msgId, payload: msg }, '*');
-
-          let transact = (e: MessageEvent) => {
-            if (e.data.pageId === PAGE_ID && e.data.msgId === msgId && e.data.type === 'reply') {
-              window.removeEventListener('message', transact, false);
-              resolve(e.data.value);
-            }
-          };
-
-          window.addEventListener('message', transact, false);
-        });
-      }
-    });
+    this.call = registerClient();
   }
 
   isConnected() {
@@ -39,7 +12,7 @@ class CasperLabsPluginHelper {
   }
 
   async sign(message: Uint8Array) {
-    return this.rpc.call<Uint8Array>('sign', message);
+    return this.call<Uint8Array>('sign', message);
   }
 }
