+++ explorer/plugin/src/lib/rpc/Provider.ts
@@ -1,48 +1,37 @@
-import { WindowPostMessageProxy } from '@ont-community/window-post-message-proxy';
 import { browser } from 'webextension-polyfill-ts';
-import { AccountApi } from './api';
-import { MethodType, Rpc } from './rpc';
+import { Rpc } from './rpc';
+import SignMessageManager from '../../background/SignMessageManager';
+
+// inject page -> content script -> background
+export function registerContentProxy(
+  logMessages = false
+) {
+  // forward messages from inpage to background
+  window.addEventListener('message', receiveMessage, false);
+
+  async function receiveMessage(event: MessageEvent) {
+    const msg = event.data;
+    if (logMessages) {
+      console.log('receive message', msg);
+    }
+    // validate message type
+    if (typeof msg !== 'object') return;
+    if (msg.type !== 'request') return;
+    msg.value = await browser.runtime.sendMessage(msg.payload);
+    msg.type = 'reply';
+    window.postMessage(msg, '*');
+  }
+}
 
 let rpc: Rpc;
 
-export function registerContentProxy(
-  {
-    logMessages = false,
-    logWarnings = false
-  }: {
-    logMessages?: boolean;
-    logWarnings?: boolean;
-  }
-) {
-  const windowPostMessageProxy = new WindowPostMessageProxy({
-    logMessages,
-    suppressWarnings: !logWarnings,
-    name: 'content-script',
-    target: 'page'
-  });
-
-  windowPostMessageProxy.addHandler({
-    handle: (msg) => browser.runtime.sendMessage(msg),
-    test: (msg) => msg.type === 'casperlabs-plugin' && msg.source === 'page'
-  });
-}
-
-export function registerProvider({ provider, logMessages }: { provider: AccountApi; logMessages: boolean }) {
+// Setup RPC server for inject page
+export function setupInjectPageAPIServer(provider: SignMessageManager, logMessages: boolean = false) {
   rpc = new Rpc({
-    source: 'background',
-    destination: 'page',
+    addListener: browser.runtime.onMessage.addListener,
     logMessages,
-    addListener: browser.runtime.onMessage.addListener
+    destination: 'page',
+    source: 'background'
   });
-
-  function checkedRegister(name: string, method: MethodType | undefined) {
-    if (method === undefined) {
-      throw new Error('DApi provider does not implement ' + name);
-    }
-
-    rpc.register(name, method);
-  }
-
-  checkedRegister('account.unlock', provider.unlock);
-  console.log("register successfully")
+  rpc.register('sign', provider.addUnsignedMessageAsync.bind(provider));
 }
