--- explorer/plugin/src/lib/rpc/api.ts
@@ -1,3 +1,0 @@
-export interface AccountApi {
-  unlock(password: string): Promise<void>;
-}
